+++ chflux/io/exceptions.py
@@ -1,0 +1,6 @@
+class ConfigException(Exception):
+    pass
+
+
+class TimestampException(Exception):
+    pass